feat(docs): 优化docs.go，给用来解析docs.yml的代码，提到parse.go
